hrms-lite-backend/requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
email-validator